requirements.txt
atproto
python-dotenv
opencv-python
# SQLite is used from the standard library
